requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
a2a-sdk==0.3.9  # Official Google A2A SDK
httpx==0.26.0  # Still needed for some HTTP operations
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.23.3